Safari-611.1.10/LayoutTests/http/tests/navigation/redirect-on-back-updates-history-item-expected.txt

main frame - has 1 onunload handler(s)
PASS

============== Back Forward List ==============
        http://127.0.0.1:8000/navigation/redirect-on-back-updates-history-item.html  **nav target**
curr->  http://127.0.0.1:8000/navigation/resources/redirect-updates-history-item-done.html  **nav target**
        http://127.0.0.1:8000/navigation/resources/redirect-on-back-updates-history-item-goback.html  **nav target**
===============================================
